Improve binary

+++ src/main/java/cn/nukkit/item/Item.java
@@ -410,7 +410,7 @@ public class Item implements Cloneable {
 
     protected Block block = null;
     protected int id;
-    protected Short meta;
+    protected Integer meta;
     private byte[] tags = new byte[0];
     private CompoundTag cachedNBT = null;
     public int count;
@@ -431,11 +431,11 @@ public class Item implements Cloneable {
 
     public Item(int id, Integer meta, int count, String name) {
         this.id = id & 0xffff;
-        this.meta = meta != null ? (short) (meta & 0xffff) : null;
+        this.meta = meta != null ? (meta & 0xffff) : null;
         this.count = count;
         this.name = name;
         if (!(this.block == null && this.id <= 0xff && Block.list[id] != null)) {
-            this.block = Block.get(this.id, Integer.valueOf(this.meta));
+            this.block = Block.get(this.id, this.meta);
             this.name = this.block.getName();
         }
     }
@@ -935,12 +935,12 @@ public class Item implements Cloneable {
         return id;
     }
 
-    public Short getDamage() {
+    public Integer getDamage() {
         return meta;
     }
 
     public void setDamage(Integer meta) {
-        this.meta = meta != null ? (short) (meta & 0xFFFF) : null;
+        this.meta = meta != null ? (meta & 0xFFFF) : null;
     }
 
     public int getMaxStackSize() {
